pyproject.toml
[project]
name = "jstapp"
version = "0.1.0"
description = "JstByte FastAPI Template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
